COMP3207/coursework/backend/pyproject.toml
[tool.poetry]
name = "kissa"
version = "0.1.0"
description = ""
authors = ["Daniel Braghis <ddb1u20@soton.ac.uk>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "0.105.0"
uvicorn = "0.24.0"
python-multipart = "^0.0.6"
passlib = "^1.7.4"
pymongo = "^4.6.1"
azure-identity = "^1.0"
python-jose = "^3.3.0"
pydantic = "^2.5.3"
email-validator = "^2.1.0.post1"
azure-keyvault-secrets = "^4.7.0"
azure-keyvault-keys = "^4.8.0"
azure-keyvault-certificates = "^4.7.0"
websockets = "^12.0"
python-dateutil = "^2.8.2"
py-dateutil = "^2.2"


[tool.poetry.group.dev.dependencies]
black = "23.12.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
